Move polyfills to utils
