+++ src/cli/config.ts
@@ -5,15 +5,16 @@ import stripJsonComments from 'strip-json-comments'
 import * as yaml from 'js-yaml'
 import { ConfigOptions } from '.'
 import yargs from 'yargs'
+import { flat } from '../utils'
 
 async function getConfigPath(dir: string, patterns: string[]): Promise<string | undefined> {
-	const files = (await Promise.all(patterns.map(async pattern => {
+	const files = flat(await Promise.all(patterns.map(async pattern => {
 		try {
 			return await glob(pattern, { absolute: true, cwd: dir, dot: true, filesOnly: true })
 		} catch {
 			return []
 		}
-	}))).flat()
+	})))
 
 	if (files.length > 0) {
 		// TODO: handle multiple config files
