+++ src/parser/document-builder.ts
@@ -2,6 +2,7 @@ import { Node, Document, ImportNode, NodeType, DiagNode, ChildContextNode, TypeN
 import { parseBindingExpression } from './compile-bindings'
 import { Diagnostic, diagnostics } from '../diagnostic'
 import { Reporting } from '../program'
+import { flat } from '../utils'
 
 // Build binding tree
 export function createDocument(filePath: string, tokens: Node[], reporting: Reporting): Document {
@@ -59,7 +60,7 @@ export function createDocument(filePath: string, tokens: Node[], reporting: Repo
 				if (node.bindings?.length) {
 					// There could be multiple data-binds on one element. Parse them all.
 					// TODO: If there are multiple binding properties on one row, emit a warning if it controls descendants (unclear semantics of multiple bindings?)
-					const bindings = node.bindings.map(bindingData => parseBindingExpression(filePath, reporting, bindingData.bindingText, bindingData.location)).flat()
+					const bindings = flat(node.bindings.map(bindingData => parseBindingExpression(filePath, reporting, bindingData.bindingText, bindingData.location)))
 					if (bindings.length) {
 						for(const b of bindings)
 							allBindings.add(b.bindingHandler.name)
@@ -82,7 +83,7 @@ export function createDocument(filePath: string, tokens: Node[], reporting: Repo
 			case NodeType.Empty: {
 				if (node.bindings?.length) {
 					// TODO: parse all binding strings
-					const bindings = node.bindings.map(bindingData => parseBindingExpression(filePath, reporting, bindingData.bindingText, bindingData.location)).flat()
+					const bindings = flat(node.bindings.map(bindingData => parseBindingExpression(filePath, reporting, bindingData.bindingText, bindingData.location)))
 					if (bindings.length) {
 						for(const b of bindings)
 							allBindings.add(b.bindingHandler.name)
